requirements.txt
# pip install -r requirements.txt, pip install sqlalchemy databases[postgresql] seperately

pandas==2.0.3
numpy==1.25.1
nltk==3.8.1
perfplot==0.10.2
matplotlib==3.7.2
psycopg2==2.9.9

# for .env 
python-dotenv==1.0.0
pydantic-settings

sqlalchemy
databases
tortoise-orm # for db
aerich # for db migration

# flask==3.0.0
fastapi==0.104.0
uvicron==0.23.2 # for fastapi

